src/downloader/requirements.txt
attrs==23.1.0 ; python_version >= "3.10" and python_version < "3.11"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "3.11"
charset-normalizer==3.1.0 ; python_version >= "3.10" and python_version < "3.11"
click-plugins==1.1.1 ; python_version >= "3.10" and python_version < "3.11"
click==8.1.3 ; python_version >= "3.10" and python_version < "3.11"
cligj==0.7.2 ; python_version >= "3.10" and python_version < "3.11"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.11" and platform_system == "Windows"
fiona==1.9.3 ; python_version >= "3.10" and python_version < "3.11"
geojson==3.0.1 ; python_version >= "3.10" and python_version < "3.11"
geomet==1.0.0 ; python_version >= "3.10" and python_version < "3.11"
geopandas==0.12.2 ; python_version >= "3.10" and python_version < "3.11"
html2text==2020.1.16 ; python_version >= "3.10" and python_version < "3.11"
idna==3.4 ; python_version >= "3.10" and python_version < "3.11"
load-dotenv==0.1.0 ; python_version >= "3.10" and python_version < "3.11"
munch==2.5.0 ; python_version >= "3.10" and python_version < "3.11"
numpy==1.23.5 ; python_version >= "3.10" and python_version < "3.11"
packaging==23.1 ; python_version >= "3.10" and python_version < "3.11"
pandas==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
pyproj==3.5.0 ; python_version >= "3.10" and python_version < "3.11"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "3.11"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "3.11"
pytz==2023.3 ; python_version >= "3.10" and python_version < "3.11"
requests==2.29.0 ; python_version >= "3.10" and python_version < "3.11"
sentinelsat==1.2.1 ; python_version >= "3.10" and python_version < "3.11"
shapely==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
six==1.16.0 ; python_version >= "3.10" and python_version < "3.11"
tqdm==4.65.0 ; python_version >= "3.10" and python_version < "3.11"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "3.11"
urllib3==1.26.15 ; python_version >= "3.10" and python_version < "3.11"
